requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
pymongo
langchain_openai
langchain_community
langchain-chroma